Crate gear_lazy_pages

Lazy pages support. In runtime data for contract wasm memory pages can be loaded in lazy manner. All pages, which is supposed to be lazy, must be mprotected before contract execution. During execution data from storage is loaded for all pages, which has been accesed. See also sys::user_signal_handler in the source code.

Restrict any page handling in signal handler more then one time. If some page will be released twice it means, that this page has been added to lazy pages more then one time during current execution. This situation may cause problems with memory data update in storage. For example: one page has no data in storage, but allocated for current program. Let’s make some action for it:

Change data in page: Default data -> Data1

Free page

Alloc page, data will Data2 (may be equal Data1).

After alloc we can set page as lazy, to identify wether page is changed after allocation. This means that we can skip page update in storage in case it wasnt changed after allocation.

Write some data in page but do not change it Data2 -> Data2. During this step signal handler writes Data2 as data for released page.

After execution we will have Data2 in page. And Data2 in released. So, nothing will be updated in storage. But program may have some significant data for next execution - so we have a bug. To avoid this we restrict double releasing. You can also check another cases in test: memory_access_cases.

Enums

Error

Functions

append_lazy_pages_infoDeprecated

Set lazy pages info and program pages prefix in global context

get_lazy_pages_numbersDeprecated

get_released_page_data

Returns data for released page

get_released_pages

Returns vec of lazy pages which has been accessed

get_wasm_mem_addr

Returns current wasm mem buffer pointer, if it’s set

get_wasm_mem_size

Returns current wasm mem buffer size, if it’s set

init⚠

Initialize lazy pages:

is_enabled

Returns whether lazy pages env is enabled

reset_context

Reset lazy pages info

set_lazy_page_infoDeprecated

Set storage key for page in global context

set_program_prefix

Set program pages prefix in storage in global context

set_wasm_mem_begin_addr

Set current wasm memory begin addr in global context

set_wasm_mem_size

Set current wasm memory size in global context