Crate gear_lazy_pages

Lazy pages support. In runtime data for contract wasm memory pages can be loaded in lazy manner. All pages, which is supposed to be lazy, must be mprotected before contract execution. During execution data from storage is loaded for all pages, which has been accesed. See also handle_sigsegv.

Structs

GetReleasedPageError

Functions

get_lazy_pages_numbers

Returns vec of not-accessed wasm lazy pages

get_released_page_old_data

Returns page data which page has in storage before execution

get_released_pages

Returns vec of lazy pages which has been accessed

init_lazy_pages⚠

Initialize lazy pages:

reset_lazy_pages_info

Reset lazy pages info

save_page_lazy_info

Save page key in storage

set_wasm_mem_begin_addr

Set current wasm memory begin addr