Crate gear_lazy_pages

Lazy-pages support. In runtime data for program Wasm memory pages can be loaded in lazy manner. All pages, which is supposed to be lazy, must be mprotected before program execution. During execution data from storage is loaded for all pages, which has been accessed and which has data in storage. See also process::process_lazy_pages, signal, host_func for more information.

Note: currently we restrict twice write signal from same page during one execution. It’s not necessary behavior, but more simple and safe.

Structs

ExceptionInfo

Enums

Error

InitError

LazyPagesError

LazyPagesVersion

Lazy-pages version.

Traits

LazyPagesStorage

UserSignalHandler

Functions

change_wasm_mem_addr_and_size

Set current wasm memory begin addr in global context

init

init_with_handler

Initialize lazy-pages for current thread.

initialize_for_program

pre_process_memory_accesses

set_lazy_pages_protection

Protect lazy pages, after they had been unprotected.

status

unset_lazy_pages_protection

Unset lazy pages read/write protections.

write_accessed_pages

Returns vec of lazy-pages which has been accessed